Add 'Gemfile' and 'Cheffile'.
Get us to a simple provisioning test where we can use Vagrant and
Chef-Solo to perform all the testing. The Cheffile just pulls in the
dependencies and adds them to the 'cookbooks' directory.

Right now because of how Vagrant works (maybe hotpatch later?) it seems
to require the roles and data_bags directories. It'd be nice to omit
them if possible.