Modifies some dependencies for the cookbook.
Properly sets the golang version (and architecture) for the vagrant
testing. Also adding some dependencies to the Cheffile for git and what
not. It seems that the doozer repository (or maybe go) requires
mercurial to be available.

Other than that modified attributes/default.rb accordingly for the
version information, installation path and repositories.